Fix missing include of compute-only driver during installed driver selection
In case the compute-only driver was installed instead of the full driver, this
plugin will not work as it will not properly recognize the compute-only driver
during installed driver selection. This commit ensures, that the compute-only
driver is also honored during installed driver selection by including the package
name for the compute-only driver in the appropriate filter.

fixes NVIDIA#13

Signed-off-by: djessich <dominik.jessich@gmail.com>